building model

+++ routes/user.js
@@ -1,11 +1,15 @@
 import { Router } from "express";
 import auth from "../utils/auth";
+import role from "../utils/role";
+import userController from "../controller/user";
 
 var router = Router();
 
-import userController from "../controller/user";
 
 router.post("/", auth.required, async (req, res, next) => {
+  if (!role.isAdmin(req.payload.username)) {
+    return res.status(403).json({ msg: "noPermission" });
+  }
   userController.createUser(req, res, next);
 });
 
@@ -14,6 +18,9 @@ router.put("/", auth.required, async (req, res, next) => {
 });
 
 router.delete("/", auth.required, async (req, res, next) => {
+  if (!role.isAdmin(req.payload.username)) {
+    return res.status(403).json({ msg: "noPermission" });
+  }
   userController.deleteUser(req, res, next);
 });
